+++ src/Executorlibs.Bilibili.Protocol/Models/Danmaku/Medal.cs
@@ -9,7 +9,7 @@ namespace Executorlibs.Bilibili.Protocol.Models.Danmaku
     /// 此类型为 Bilibili 直播平台 专用<para/>
     /// 继承自 <see cref="IMedal{TUserId}"/>
     /// </remarks>
-    public interface IMedal : IMedal<int>
+    public interface IMedal : IMedal<long>
     {
         /// <summary>
         /// 前置小图标
@@ -20,7 +20,7 @@ namespace Executorlibs.Bilibili.Protocol.Models.Danmaku
     /// <summary>
     /// 表示勋章信息的类
     /// </summary>
-    public class Medal : Medal<int>, IMedal
+    public class Medal : Medal<long>, IMedal
     {
         /// <inheritdoc/>
         public int? Badge { get; set; }
